Trait sp_state_machine::backend::Backend

pub trait Backend<H: Hasher>: Debug {
    type Error: Error;
    type Transaction: Consolidate + Default;
    type TrieBackendStorage: TrieBackendStorage<H>;
    fn storage(&self, key: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>;

    fn child_storage(
        &self, 
        storage_key: &[u8], 
        child_info: ChildInfo, 
        key: &[u8]
    ) -> Result<Option<Vec<u8>>, Self::Error>;

    fn next_storage_key(
        &self, 
        key: &[u8]
    ) -> Result<Option<Vec<u8>>, Self::Error>;

    fn next_child_storage_key(
        &self, 
        storage_key: &[u8], 
        child_info: ChildInfo, 
        key: &[u8]
    ) -> Result<Option<Vec<u8>>, Self::Error>;

    fn for_keys_in_child_storage<F: FnMut(&[u8])>(
        &self, 
        storage_key: &[u8], 
        child_info: ChildInfo, 
        f: F
    );

    fn for_key_values_with_prefix<F: FnMut(&[u8], &[u8])>(
        &self, 
        prefix: &[u8], 
        f: F
    );

    fn for_child_keys_with_prefix<F: FnMut(&[u8])>(
        &self, 
        storage_key: &[u8], 
        child_info: ChildInfo, 
        prefix: &[u8], 
        f: F
    );

    fn storage_root<I>(&self, delta: I) -> (H::Out, Self::Transaction)
    where
        I: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,
        H::Out: Ord;

    fn child_storage_root<I>(
        &self, 
        storage_key: &[u8], 
        child_info: ChildInfo, 
        delta: I
    ) -> (H::Out, bool, Self::Transaction)
    where
        I: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,
        H::Out: Ord;

    fn pairs(&self) -> Vec<(Vec<u8>, Vec<u8>)>;

    fn storage_hash(&self, key: &[u8]) -> Result<Option<H::Out>, Self::Error> { ... }

    fn child_storage_hash(
        &self, 
        storage_key: &[u8], 
        child_info: ChildInfo, 
        key: &[u8]
    ) -> Result<Option<H::Out>, Self::Error> { ... }

    fn exists_storage(&self, key: &[u8]) -> Result<bool, Self::Error> { ... }

    fn exists_child_storage(
        &self, 
        storage_key: &[u8], 
        child_info: ChildInfo, 
        key: &[u8]
    ) -> Result<bool, Self::Error> { ... }

    fn for_keys_with_prefix<F: FnMut(&[u8])>(&self, prefix: &[u8], f: F) { ... }

    fn keys(&self, prefix: &[u8]) -> Vec<Vec<u8>> { ... }

    fn child_keys(
        &self, 
        storage_key: &[u8], 
        child_info: ChildInfo, 
        prefix: &[u8]
    ) -> Vec<Vec<u8>> { ... }

    fn as_trie_backend(
        &mut self
    ) -> Option<&TrieBackend<Self::TrieBackendStorage, H>> { ... }

    fn full_storage_root<I1, I2i, I2>(
        &self, 
        delta: I1, 
        child_deltas: I2
    ) -> (H::Out, Self::Transaction)
    where
        I1: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,
        I2i: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,
        I2: IntoIterator<Item = (Vec<u8>, I2i, OwnedChildInfo)>,
        H::Out: Ord + Encode,
    { ... }
}

A state backend is used to read state data and can have changes committed to it.

The clone operation (if implemented) should be cheap.

Associated Types

type Error: Error

An error type when fetching data is not possible.

type Transaction: Consolidate + Default

Storage changes to be applied if committing

type TrieBackendStorage: TrieBackendStorage<H>

Type of trie backend storage.

Loading content...
Required methods

fn storage(&self, key: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>

Get keyed storage or None if there is nothing associated.

fn child_storage(
&self,
storage_key: &[u8],
child_info: ChildInfo,
key: &[u8]
) -> Result<Option<Vec<u8>>, Self::Error>

Get keyed child storage or None if there is nothing associated.

fn next_storage_key(&self, key: &[u8]) -> Result<Option<Vec<u8>>, Self::Error>

Return the next key in storage in lexicographic order or None if there is no value.

fn next_child_storage_key(
&self,
storage_key: &[u8],
child_info: ChildInfo,
key: &[u8]
) -> Result<Option<Vec<u8>>, Self::Error>

Return the next key in child storage in lexicographic order or None if there is no value.

fn for_keys_in_child_storage<F: FnMut(&[u8])>(
&self,
storage_key: &[u8],
child_info: ChildInfo,
f: F
)

Retrieve all entries keys of child storage and call f for each of those keys.

fn for_key_values_with_prefix<F: FnMut(&[u8], &[u8])>(
&self,
prefix: &[u8],
f: F
)

Retrieve all entries keys and values of which start with the given prefix and call f for each of those keys.

fn for_child_keys_with_prefix<F: FnMut(&[u8])>(
&self,
storage_key: &[u8],
child_info: ChildInfo,
prefix: &[u8],
f: F
)

Retrieve all child entries keys which start with the given prefix and call f for each of those keys.

fn storage_root<I>(&self, delta: I) -> (H::Out, Self::Transaction) where
I: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,
H::Out: Ord,

Calculate the storage root, with given delta over what is already stored in the backend, and produce a "transaction" that can be used to commit. Does not include child storage updates.

fn child_storage_root<I>(
&self,
storage_key: &[u8],
child_info: ChildInfo,
delta: I
) -> (H::Out, bool, Self::Transaction) where
I: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,
H::Out: Ord,

Calculate the child storage root, with given delta over what is already stored in the backend, and produce a "transaction" that can be used to commit. The second argument is true if child storage root equals default storage root.

fn pairs(&self) -> Vec<(Vec<u8>, Vec<u8>)>

Get all key/value pairs into a Vec.

Loading content...
Provided methods

fn storage_hash(&self, key: &[u8]) -> Result<Option<H::Out>, Self::Error>

Get keyed storage value hash or None if there is nothing associated.

fn child_storage_hash(
&self,
storage_key: &[u8],
child_info: ChildInfo,
key: &[u8]
) -> Result<Option<H::Out>, Self::Error>

Get child keyed storage value hash or None if there is nothing associated.

fn exists_storage(&self, key: &[u8]) -> Result<bool, Self::Error>

true if a key exists in storage.

fn exists_child_storage(
&self,
storage_key: &[u8],
child_info: ChildInfo,
key: &[u8]
) -> Result<bool, Self::Error>

true if a key exists in child storage.

fn for_keys_with_prefix<F: FnMut(&[u8])>(&self, prefix: &[u8], f: F)

Retrieve all entries keys which start with the given prefix and call f for each of those keys.

fn keys(&self, prefix: &[u8]) -> Vec<Vec<u8>>

Get all keys with given prefix

fn child_keys(
&self,
storage_key: &[u8],
child_info: ChildInfo,
prefix: &[u8]
) -> Vec<Vec<u8>>

Get all keys of child storage with given prefix

fn as_trie_backend(
&mut self
) -> Option<&TrieBackend<Self::TrieBackendStorage, H>>

Try convert into trie backend.

fn full_storage_root<I1, I2i, I2>(
&self,
delta: I1,
child_deltas: I2
) -> (H::Out, Self::Transaction) where
I1: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,
I2i: IntoIterator<Item = (Vec<u8>, Option<Vec<u8>>)>,
I2: IntoIterator<Item = (Vec<u8>, I2i, OwnedChildInfo)>,
H::Out: Ord + Encode,

Calculate the storage root, with given delta over what is already stored in the backend, and produce a "transaction" that can be used to commit. Does include child storage updates.

Loading content...