Enum sp_state_machine::ExecutionError

pub enum ExecutionError {
    Backend(DefaultError),
    CodeEntryDoesNotExist,
    UnableToGenerateProof,
    InvalidProof,
}

Externalities Error.

Externalities are not really allowed to have errors, since itโs assumed that dependent code would not be executed unless externalities were available. This is included for completeness, and as a transition away from the pre-existing framework.

Variants

Backend(DefaultError)

CodeEntryDoesNotExist

UnableToGenerateProof

InvalidProof