Function sp_state_machine::oldest_non_pruned_changes_trie

pub fn oldest_non_pruned_changes_trie<Number: BlockNumber>(
    config: &ChangesTrieConfiguration, 
    min_blocks_to_keep: Number, 
    best_finalized_block: Number
) -> Number

Get number of oldest block for which changes trie is not pruned given changes trie configuration, pruning parameter and number of best finalized block.