Function sp_state_machine::prove_range_read_with_child_with_size

pub fn prove_range_read_with_child_with_size<B, H>(
    backend: B,
    size_limit: usize,
    start_at: &[Vec<u8>],
) -> Result<(StorageProof, u32), Box<dyn Error>>
where
    B: AsTrieBackend<H>,
    H: Hasher,
    H::Out: Ord + Codec,

Generate range storage read proof, with child tries content. A size limit is applied to the proof with the exception that start_at and its following element are always part of the proof. If a key different than start_at is a child trie root, the child trie content will be included in the proof.