sp_state_machine

Function read_range_proof_check_on_proving_backend

pub fn read_range_proof_check_on_proving_backend<H>(
    proving_backend: &TrieBackend<MemoryDB<H>, H>,
    child_info: Option<&ChildInfo>,
    prefix: Option<&[u8]>,
    count: Option<u32>,
    start_at: Option<&[u8]>,
) -> Result<(Vec<(Vec<u8>, Vec<u8>)>, bool), Box<dyn Error>>
where
    H: Hasher,
    H::Out: Ord + Codec,

Check storage range proof on pre-created proving backend.

Returns a vector with the read key => value pairs and a bool that is set to true when all key => value pairs could be read and no more are left.