Struct sp_state_machine::ChangesTrieState

pub struct ChangesTrieState<'a, H, Number> {
    pub config: ChangesTrieConfiguration,
    pub zero: Number,
    pub storage: &'a dyn Storage<H, Number>,
}

Changes tries state at some block.

Fields
config: ChangesTrieConfiguration

Configuration that is active at given block.

zero: Number

Configuration activation block number. Zero if it is the first configuration on the chain, or number of the block that have emit NewConfiguration signal (thus activating configuration starting from the next block).

storage: &'a dyn Storage<H, Number>

Underlying changes tries storage reference.

Implementations

impl<'a, H, Number> State<'a, H, Number>

pub fn new(
config: ChangesTrieConfiguration,
zero: Number,
storage: &'a dyn Storage<H, Number>
) -> Self

Create state with given config and storage.