Struct sp_state_machine::ProvingBackend

pub struct ProvingBackend<'a, S: 'a + TrieBackendStorage<H>, H: 'a + Hasher>(_);

Patricia trie-based backend which also tracks all touched storage trie values. These can be sent to remote node and used as a proof of execution.

Implementations

impl<'a, S: 'a + TrieBackendStorage<H>, H: 'a + Hasher> ProvingBackend<'a, S, H> where
H::Out: Codec,

pub fn new(backend: &'a TrieBackend<S, H>) -> Self

Create new proving backend.

pub fn new_with_recorder(
backend: &'a TrieBackend<S, H>,
proof_recorder: ProofRecorder<H>
) -> Self

Create new proving backend with the given recorder.

pub fn extract_proof(&self) -> StorageProof

Extracting the gathered unordered proof.