Struct sp_state_machine::StorageProof

pub struct StorageProof { /* private fields */ }

A proof that some set of key-value pairs are included in the storage trie. The proof contains the storage values so that the partial storage backend can be reconstructed by a verifier that does not already have access to the key-value pairs.

The proof consists of the set of serialized nodes in the storage trie accessed when looking up the keys covered by the proof. Verifying the proof requires constructing the partial trie from the serialized nodes and performing the key lookups.

Implementations

impl StorageProof

pub fn new( trie_nodes: impl IntoIterator<Item = Vec<u8, Global>> ) -> StorageProof

Constructs a storage proof from a subset of encoded trie nodes in a storage backend.

pub fn empty() -> StorageProof

Returns a new empty proof.

An empty proof is capable of only proving trivial statements (ie. that an empty set of key-value pairs exist in storage).

pub fn is_empty(&self) -> bool

Returns whether this is an empty proof.

pub fn into_iter_nodes(self) -> impl DoubleEndedIterator

Convert into an iterator over encoded trie nodes in lexicographical order constructed from the proof.

pub fn iter_nodes(&self) -> impl DoubleEndedIterator

Create an iterator over encoded trie nodes in lexicographical order constructed from the proof.

pub fn into_nodes(self) -> BTreeSet<Vec<u8, Global>, Global>

Convert into plain node vector.

pub fn into_memory_db<H>(self) -> MemoryDB<H, HashKey<H>, Vec<u8, Global>>where H: Hasher,

Creates a MemoryDB from Self.

pub fn to_memory_db<H>(&self) -> MemoryDB<H, HashKey<H>, Vec<u8, Global>>where H: Hasher,

Creates a MemoryDB from Self reference.

pub fn merge(proofs: impl IntoIterator<Item = StorageProof>) -> StorageProof

Merges multiple storage proofs covering potentially different sets of keys into one proof covering all keys. The merged proof output may be smaller than the aggregate size of the input proofs due to deduplication of trie nodes.

pub fn into_compact_proof<H>( self, root: <H as Hasher>::Out ) -> Result<CompactProof, Error<<H as Hasher>::Out, Error<<H as Hasher>::Out>>>where H: Hasher,

Encode as a compact proof with default trie layout.

pub fn to_compact_proof<H>( &self, root: <H as Hasher>::Out ) -> Result<CompactProof, Error<<H as Hasher>::Out, Error<<H as Hasher>::Out>>>where H: Hasher,

Encode as a compact proof with default trie layout.

pub fn encoded_compact_size<H>(self, root: <H as Hasher>::Out) -> Option<usize>where H: Hasher,

Returns the estimated encoded size of the compact proof.

Running this operation is a slow operation (build the whole compact proof) and should only be in non sensitive path.

Return None on error.