sp_state_machine

Module backend

State machine backends. These manage the code and storage of contracts.

Structs

BackendRuntimeCode

Wrapper to create a RuntimeCode from a type that implements Backend.

IterArgs

A struct containing arguments for iterating over the storage.

KeysIter

An iterator over storage keys.

PairsIter

An iterator over storage keys and values.

Traits

AsTrieBackend

Something that can be converted into a TrieBackend.

Backend

A state backend is used to read state data and can have changes committed to it.

StorageIterator

A trait for a raw storage iterator.

Type Aliases

BackendTransaction

The transaction type used by Backend.