Struct sp_state_machine::BasicExternalities

pub struct BasicExternalities { /* fields omitted */ }

Simple Map-based Externalities impl.

Implementations

impl BasicExternalities

pub fn new(inner: Storage) -> Self

Create a new instance of BasicExternalities

pub fn new_empty() -> Self

New basic externalities with empty storage.

pub fn insert(&mut self, k: StorageKey, v: StorageValue) -> Option<StorageValue>

Insert key/value

pub fn into_storages(self) -> Storage

Consume self and returns inner storages

pub fn execute_with_storage<R>(
storage: &mut Storage,
f: impl FnOnce() -> R
) -> R

Execute the given closure f with the externalities set and initialized with storage.

Returns the result of the closure and updates storage with all changes.

pub fn execute_with<R>(&mut self, f: impl FnOnce() -> R) -> R

Execute the given closure while self is set as externalities.

Returns the result of the given closure.

pub fn extensions(&mut self) -> &mut Extensions

List of active extensions.

pub fn register_extension(&mut self, ext: impl Extension)

Register an extension.