Struct sp_state_machine::BasicExternalities

pub struct BasicExternalities { /* fields omitted */ }

Simple Map-based Externalities impl.

Methods

impl BasicExternalities

pub fn new(inner: Storage) -> Self

Create a new instance of BasicExternalities

pub fn insert(&mut self, k: Vec<u8>, v: Vec<u8>) -> Option<Vec<u8>>

Insert key/value

pub fn into_storages(self) -> Storage

Consume self and returns inner storages

pub fn execute_with_storage<R>(
storage: &mut Storage,
f: impl FnOnce() -> R
) -> R

Execute the given closure f with the externalities set and initialized with storage.

Returns the result of the closure and updates storage with all changes.

pub fn execute_with<R>(&mut self, f: impl FnOnce() -> R) -> R

Execute the given closure while self is set as externalities.

Returns the result of the given closure.