Struct sp_state_machine::TestExternalities

pub struct TestExternalities<H: Hasher, N: ChangesTrieBlockNumber = u64> where
    H::Out: Codec + Ord,  { /* fields omitted */ }

Simple HashMap-based Externalities impl.

Implementations

impl<H: Hasher, N: ChangesTrieBlockNumber> TestExternalities<H, N> where
H::Out: Ord + 'static + Codec,

pub fn ext(&mut self) -> Ext<'_, H, N, InMemoryBackend<H>>

Get externalities implementation.

pub fn new(storage: Storage) -> Self

Create a new instance of TestExternalities with storage.

pub fn new_empty() -> Self

New empty test externalities.

pub fn new_with_code(code: &[u8], storage: Storage) -> Self

Create a new instance of TestExternalities with code and storage.

pub fn persist_offchain_overlay(&mut self)

Move offchain changes from overlay to the persistent store.

pub fn offchain_db(&self) -> TestPersistentOffchainDB

A shared reference type around the offchain worker storage.

pub fn insert(&mut self, k: StorageKey, v: StorageValue)

Insert key/value into backend

pub fn register_extension<E: Any + Extension>(&mut self, ext: E)

Registers the given extension for this instance.

pub fn changes_trie_storage(&mut self) -> &mut ChangesTrieInMemoryStorage<H, N>

Get mutable reference to changes trie storage.

pub fn commit_all(&self) -> InMemoryBackend<H>

Return a new backend with all pending value.

pub fn execute_with<R>(&mut self, execute: impl FnOnce() -> R) -> R

Execute the given closure while self is set as externalities.

Returns the result of the given closure.