Struct sp_state_machine::TestExternalities

pub struct TestExternalities<H: Hasher> where
    H::Out: Codec + Ord,  {
    pub backend: InMemoryBackend<H>,
    pub extensions: Extensions,
    pub state_version: StateVersion,
    /* private fields */
}

Simple HashMap-based Externalities impl.

Fields
backend: InMemoryBackend<H>

Storage backend.

extensions: Extensions

Extensions.

state_version: StateVersion

State version to use during tests.

Implementations

impl<H: Hasher> TestExternalities<H> where
H::Out: Ord + 'static + Codec,

pub fn ext(&mut self) -> Ext<'_, H, InMemoryBackend<H>>

Get externalities implementation.

pub fn new(storage: Storage) -> Self

Create a new instance of TestExternalities with storage.

pub fn new_with_state_version(
storage: Storage,
state_version: StateVersion
) -> Self

Create a new instance of TestExternalities with storage for a given state version.

pub fn new_empty() -> Self

New empty test externalities.

pub fn new_with_code(code: &[u8], storage: Storage) -> Self

Create a new instance of TestExternalities with code and storage.

pub fn new_with_code_and_state(
code: &[u8],
storage: Storage,
state_version: StateVersion
) -> Self

Create a new instance of TestExternalities with code and storage for a given state version.

pub fn overlayed_changes(&self) -> &OverlayedChanges

Returns the overlayed changes.

pub fn persist_offchain_overlay(&mut self)

Move offchain changes from overlay to the persistent store.

pub fn offchain_db(&self) -> TestPersistentOffchainDB

A shared reference type around the offchain worker storage.

pub fn insert(&mut self, k: StorageKey, v: StorageValue)

Insert key/value into backend

pub fn insert_child(&mut self, c: ChildInfo, k: StorageKey, v: StorageValue)

Insert key/value into backend.

This only supports inserting keys in child tries.

pub fn register_extension<E: Any + Extension>(&mut self, ext: E)

Registers the given extension for this instance.

pub fn as_backend(&self) -> InMemoryBackend<H>

Return a new backend with all pending changes.

In contrast to commit_all this will not panic if there are open transactions.

pub fn commit_all(&mut self) -> Result<(), String>

Commit all pending changes to the underlying backend.

Panic

This will panic if there are still open transactions.

pub fn execute_with<R>(&mut self, execute: impl FnOnce() -> R) -> R

Execute the given closure while self is set as externalities.

Returns the result of the given closure.

pub fn execute_and_prove<R>(
&mut self,
execute: impl FnOnce() -> R
) -> (R, StorageProof)

Execute the given closure while self, with proving_backend as backend, is set as externalities.

This implementation will wipe the proof recorded in between calls. Consecutive calls will get their own proof from scratch.

pub fn execute_with_safe<R>(
&mut self,
f: impl FnOnce() -> R + UnwindSafe
) -> Result<R, String>

Execute the given closure while self is set as externalities.

Returns the result of the given closure, if no panics occurred. Otherwise, returns Err.