Struct sp_state_machine::TrieBackend

pub struct TrieBackend<S: TrieBackendStorage<H>, H: Hasher, C = LocalTrieCache<H>> { /* private fields */ }

Patricia trie-based backend. Transaction type is an overlay of changes to commit.

Implementations

impl<H: Hasher, KF> TrieBackend<GenericMemoryDB<H, KF>, H>where
H::Out: Codec + Ord,
KF: KeyFunction<H> + Send + Sync,

pub fn update<T: IntoIterator<Item = (Option<ChildInfo>, StorageCollection)>>(
&self,
changes: T,
state_version: StateVersion
) -> Self

Copy the state, with applied updates

pub fn insert<T: IntoIterator<Item = (Option<ChildInfo>, StorageCollection)>>(
&mut self,
changes: T,
state_version: StateVersion
)

Insert values into backend trie.

pub fn update_backend(
&self,
root: H::Out,
changes: GenericMemoryDB<H, KF>
) -> Self

Merge trie nodes into this backend.

pub fn apply_transaction(
&mut self,
root: H::Out,
transaction: GenericMemoryDB<H, KF>
)

Apply the given transaction to this backend and set the root to the given value.

pub fn eq(&self, other: &Self) -> bool

Compare with another in-memory backend.

impl<S: TrieBackendStorage<H>, H: Hasher, C: AsLocalTrieCache<H> + Send + Sync> TrieBackend<S, H, C>where
H::Out: Codec,

pub fn essence(&self) -> &TrieBackendEssence<S, H, C>

Get backend essence reference.

pub fn backend_storage(&self) -> &S

Get backend storage reference.

pub fn root(&self) -> &H::Out

Get trie root.

pub fn into_storage(self) -> S

Consumes self and returns underlying storage.

pub fn extract_proof(self) -> Option<StorageProof>

Extract the StorageProof.

This only returns Some when there was a recorder set.