Struct sp_state_machine::TrieBackend

pub struct TrieBackend<S: TrieBackendStorage<H>, H: Hasher> { /* fields omitted */ }

Patricia trie-based backend. Transaction type is an overlay of changes to commit.

Implementations

impl<H: Hasher> TrieBackend<MemoryDB<H>, H> where
H::Out: Codec + Ord,

pub fn update<T: IntoIterator<Item = (Option<ChildInfo>, StorageCollection)>>(
&self,
changes: T
) -> Self

Copy the state, with applied updates

pub fn insert<T: IntoIterator<Item = (Option<ChildInfo>, StorageCollection)>>(
&mut self,
changes: T
)

Insert values into backend trie.

pub fn update_backend(&self, root: H::Out, changes: MemoryDB<H>) -> Self

Merge trie nodes into this backend.

pub fn apply_transaction(&mut self, root: H::Out, transaction: MemoryDB<H>)

Apply the given transaction to this backend and set the root to the given value.

pub fn eq(&self, other: &Self) -> bool

Compare with another in-memory backend.

impl<S: TrieBackendStorage<H>, H: Hasher> TrieBackend<S, H> where
H::Out: Codec,

pub fn new(storage: S, root: H::Out) -> Self

Create new trie-based backend.

pub fn essence(&self) -> &TrieBackendEssence<S, H>

Get backend essence reference.

pub fn backend_storage(&self) -> &S

Get backend storage reference.

pub fn backend_storage_mut(&mut self) -> &mut S

Get backend storage reference.

pub fn root(&self) -> &H::Out

Get trie root.

pub fn into_storage(self) -> S

Consumes self and returns underlying storage.